Added

+++ src/driverapi/plugins/memrecorder/MemRecorder.cpp
@@ -694,7 +694,7 @@ extern "C" {
 	}
 
 	void DIOGENES_REC_GLIBFREE(void * addr) {
-		POSTPROCESS_GNUFREE(addr);
+		POSTPROCESS_GNUFREE((uint64_t)addr);
 		DIOGENES_libcfree_wrapper(addr);
 	}
 /*	void DIOGENES_REC_GLIBMALLOC_PRE(size_t size) {
